chore: move admin account settings
