+++ web-client/src/pages/Admin/AccountSettings/components/AccountProfileForm.tsx
@@ -2,12 +2,12 @@ import { useState, useRef } from "react";
 import { useNavigate } from "react-router-dom";
 import { FieldValues, SubmitHandler, useForm } from "react-hook-form";
 
-import { TUserResData } from "../../../types/user";
+import { TUserResData } from "../../../../types/user";
 
 import {
   useGetUserByTokenQuery,
   useUpdateProfileMutation,
-} from "../../../services/accountQuery";
+} from "../../../../services/accountQuery";
 
 import { MdClose } from "react-icons/md";
 
@@ -15,7 +15,7 @@ import "cropperjs/dist/cropper.css";
 import moment from "moment";
 
 import AccountProfilePicture from "./AccountProfilePicture";
-import AddressField from "../../../components/AddressSelector/AddressField";
+import AddressField from "../../../../components/AddressSelector/AddressField";
 
 const AccountProfileForm = () => {
   const navigate = useNavigate();
